refactor: move main method and refactor BaseBall
